🎧 extended options, bundled 3 themes and enabled theme selection
