+++ apps/simple/static/lite.html
@@ -38,9 +38,6 @@
     />
     <script src="index.js"></script>
     <style>
-      .CodeMirror {
-        max-width: 900px;
-      }
       pre.code {
         background-color: #eee;
         padding: 16px;
