+++ packages/thebe/src/types.ts
@@ -12,6 +12,28 @@ import type { Options } from './options';
 import type { KernelStatus } from './status';
 import type { ThebeLiteGlobal } from 'thebe-lite';
 
+export interface CellDOMPlaceholder {
+  id: string;
+  placeholders: {
+    source: Element;
+    output?: Element;
+  };
+}
+
+export type CellDOMItem = CellDOMPlaceholder & {
+  ui: {
+    cell: Element;
+    editor: Element;
+    output?: Element;
+    buttons: {
+      run?: Element;
+      runAll?: Element;
+      restart?: Element;
+      restartAll?: Element;
+    };
+  };
+};
+
 export interface ThebeGlobal {
   mountStatusWidget: () => void;
   mountActivateWidget: () => void;
